Remove strategy for now
