+++ postgres/eventstore/options.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 
 	"github.com/go-estoria/estoria"
-	"github.com/go-estoria/estoria-contrib/postgres/outbox"
 )
 
 type EventStoreOption func(*EventStore) error
@@ -23,22 +22,22 @@ func WithLogger(logger estoria.Logger) EventStoreOption {
 	}
 }
 
-// WithOutbox sets the outbox to use for the event store.
-func WithOutbox(outbox *outbox.Outbox) EventStoreOption {
-	return func(s *EventStore) error {
-		s.AddTransactionalHooks(outbox.HandleEvents)
-		return nil
-	}
-}
+// // WithOutbox sets the outbox to use for the event store.
+// func WithOutbox(outbox *outbox.Outbox) EventStoreOption {
+// 	return func(s *EventStore) error {
+// 		s.AddTransactionalHooks(outbox.HandleEvents)
+// 		return nil
+// 	}
+// }
 
-// WithStrategy sets a custom read/write strategy for the event store.
-// The provided strategy must implement both ReadStrategy and WriteStrategy.
-func WithStrategy(strat Strategy) EventStoreOption {
+// WithTableName sets a custom table name for the event store.
+// The table name must be a valid SQL identifier.
+func WithTableName(tableName string) EventStoreOption {
 	return func(s *EventStore) error {
-		if strat == nil {
-			return errors.New("strategy cannot be nil")
+		if tableName == "" {
+			return errors.New("table name cannot be empty")
 		}
-		s.strategy = strat
+		s.tableName = tableName
 		return nil
 	}
 }
