+++ postgres/outbox/outbox.go
@@ -1,13 +1,24 @@
 package outbox
 
 import (
+	"context"
+	"database/sql"
 	"fmt"
-
-	"github.com/go-estoria/estoria-contrib/postgres/eventstore/strategy"
 
 	"github.com/go-estoria/estoria"
 	"github.com/go-estoria/estoria/eventstore"
 )
+
+// Transaction is an interface for executing queries within a transactional context.
+// It abstracts *sql.Tx by exposing prepare, query, and exec methods, plus commit/rollback.
+type Transaction interface {
+	Prepare(query string) (*sql.Stmt, error)
+	QueryRowContext(ctx context.Context, query string, args ...any) *sql.Row
+	QueryContext(ctx context.Context, query string, args ...any) (*sql.Rows, error)
+	ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error)
+	Commit() error
+	Rollback() error
+}
 
 type Outbox struct {
 	table string
@@ -24,7 +35,7 @@ func New(table string) (*Outbox, error) {
 }
 
 // HandleEvents inserts emitted events into the outbox table using the given transactional context.
-func (o *Outbox) HandleEvents(tx strategy.Transaction, events []*eventstore.Event) error {
+func (o *Outbox) HandleEvents(tx Transaction, events []*eventstore.Event) error {
 	o.log.Debug("inserting events into outbox", "tx", "inherited", "events", len(events))
 
 	statement := fmt.Sprintf(`
